+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * [System] Convert remaining functions in System to rely on char data
 
 * Soft deprecations (no warnings emitted)
+  * [Application] use Application.Behaviour is deprecated in favor of use Application
   * [Mix] `Mix.project/0` is deprecated in favor of `Mix.Project.config/0`
   * [Process] `Process.spawn/1`, `Process.spawn/3`, `Process.spawn_link/1`, `Process.spawn_link/3`, `Process.spawn_monitor/1`, `Process.spawn_monitor/3`, `Process.send/2` and `Process.self/0` are deprecated in favor of the ones in Kernel